refactor: Fragment 애서 사용하는 ViewModel 분리
뒤로가기 등에서 데어트를 지우기 위함.
