+++ data/src/main/java/com/yapp/bol/data/remote/GroupApi.kt
@@ -9,6 +9,7 @@ import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
 import retrofit2.http.Path
+import retrofit2.http.Query
 
 interface GroupApi {
